requirements.txt
PyQt5>=5.15
mido>=1.3
python-rtmidi
pyserial>=3.5
pyFluidSynth>=1.3
